+++ pkg/apis/go.mod
@@ -3,7 +3,7 @@ module github.com/grafana/grafana/pkg/apis
 go 1.23.7
 
 require (
-	github.com/grafana/authlib/types v0.0.0-20250224151205-5ef97131cc82
+	github.com/grafana/authlib/types v0.0.0-20250314102521-a77865c746c0
 	github.com/grafana/grafana-plugin-sdk-go v0.272.0
 	github.com/grafana/grafana/pkg/apimachinery v0.0.0-20250311185328-85b0b47efdd8
 	github.com/spyzhov/ajson v0.9.6
@@ -103,9 +103,9 @@ require (
 	golang.org/x/exp v0.0.0-20240909161429-701f63a606c0 // indirect
 	golang.org/x/mod v0.23.0 // indirect
 	golang.org/x/net v0.36.0 // indirect
-	golang.org/x/sync v0.11.0 // indirect
-	golang.org/x/sys v0.30.0 // indirect
-	golang.org/x/text v0.22.0 // indirect
+	golang.org/x/sync v0.12.0 // indirect
+	golang.org/x/sys v0.31.0 // indirect
+	golang.org/x/text v0.23.0 // indirect
 	golang.org/x/tools v0.30.0 // indirect
 	golang.org/x/xerrors v0.0.0-20240903120638-7835f813f4da // indirect
 	google.golang.org/genproto/googleapis/api v0.0.0-20250218202821-56aae31c358a // indirect
